Extract more `PureSvg*Utils` files
