+++ DevToolkt/pureDom/src/jvmMain/kotlin/dev/toolkt/dom/pure/svg/PureSvgGraphicsElement.kt
@@ -1,0 +1,9 @@
+package dev.toolkt.dom.pure.svg
+
+import dev.toolkt.geometry.transformations.Transformation
+
+abstract class PureSvgGraphicsElement : PureSvgElement() {
+    abstract fun flatten(
+        baseTransformation: Transformation,
+    ): List<PureSvgShape>
+}
