+++ DevToolkt/pureDom/src/jvmMain/kotlin/dev/toolkt/dom/pure/svg/PureSvgRootUtils.kt
@@ -1,5 +1,8 @@
 package dev.toolkt.dom.pure.svg
 
+import dev.toolkt.dom.pure.PureDimension
+import dev.toolkt.dom.pure.utils.xml.childElements
+import dev.toolkt.dom.pure.utils.xml.getAttributeOrNull
 import dev.toolkt.dom.pure.utils.xml.svg.MinimalCssContext
 import dev.toolkt.dom.pure.utils.xml.svg.SVGDOMImplementationUtils
 import dev.toolkt.dom.pure.utils.xml.svg.createSvgDocument
@@ -30,5 +33,28 @@ fun PureSvgRoot.toSvgDocument(): SVGDocument = svgDomImplementation.createSvgDoc
     setup(
         document = this,
         root = documentSvgElement,
+    )
+}
+
+fun SVGDocument.toPure(): PureSvgRoot {
+    val widthString =
+        documentSvgElement.getAttributeOrNull("width") ?: throw IllegalArgumentException("Width is not set")
+    val heightString =
+        documentSvgElement.getAttributeOrNull("height") ?: throw IllegalArgumentException("Height is not set")
+
+    val width = PureDimension.parse(widthString)
+    val height = PureDimension.parse(heightString)
+
+    val viewBox = documentSvgElement.getAttributeOrNull("viewBox")?.let { viewBoxString ->
+        PureSvgRoot.ViewBox.parse(viewBoxString)
+    }
+
+    return PureSvgRoot(
+        width = width,
+        height = height,
+        viewBox = viewBox,
+        graphicsElements = documentSvgElement.childElements.mapNotNull {
+            it.toSvgGraphicsElements()
+        },
     )
 }
